move upgrade handlers to next-upgrade
